Crate win_nfd

Re-exports

pub use self::c_wide_string::CWideStr;

pub use self::c_wide_string::CWideString;

pub use self::c_wide_string::NulError;

pub use self::fileapi::get_full_path_name;

pub use self::shobjidl::DisplayNameType;

pub use self::shobjidl::FileDialog;

pub use self::shobjidl::FileFilters;

pub use self::shobjidl::FileOpenDialog;

pub use self::shobjidl::FileSaveDialog;

pub use self::shobjidl::ModalWindow;

pub use self::shobjidl::ShellItem;

Modules

c_wide_string

fileapi

shobjidl

Structs

CoTaskMemWideString

A Wide String allocated with CoTaskMemAlloc.

FileOpenDialogBuilder

Builder for a FileOpenDialog

FileSaveDialogBuilder

Builder for a FileSaveDialog

HResult

A wrapper for a windows HRESULT.

Enums

NfdError

An error that may occur during the use of a file dialog

Functions

nfd_open

Default nfd open dialog. Look at this functions impl and write your own if you need more control

nfd_open_builder

Shothand for FileOpenDialogBuilder::new().init_com()

nfd_save

Default nfd save dialog. Look at this functions impl and write your own if you need more control

nfd_save_builder

Shothand for FileSaveDialogBuilder::new().init_com()