Trait win_nfd::c_wide_string::IntoWide

pub trait IntoWide {
    fn into_wide(self) -> Vec<u16>;
}

Implemented for types that can be converted into wide types

Required methods

fn into_wide(self) -> Vec<u16>

Convert this into a vec of wide chars

Implementors should reserve 1 extra element of space for the nul terminator. The vec may contain nul elements, but they will be rejected in some places in the api.